Extract utils for line and polygon element
